refactor: move getConfigXXXX function to config.ts
